Fix decimal rounding for BridgeInputBox "Max" balance button
These changes fix-up a few issues with the "Max" button on the Bridge
page.

The token's decimals was calculated from the readable value rather than
relying on the decimals value declared in the token's contract. This was
causing rounding errors when the decimal was lead or trailed with zeros.

The balance was always rounded to 8 decimals; however, some tokens have
a smaller number of decimals. This would cause an error to be thrown.
The balance is now rounded to eight or less decimals.